feat: Add introSeen and assetsLoaded to AppStore
This commit adds the `introSeen` and `assetsLoaded` properties to the `AppStore` interface in the `use-app-store.ts` file. It also includes the corresponding `setIntroSeen` and `setAssetsLoaded` methods for updating these properties. This change allows for tracking whether the intro has been seen and whether the assets have been loaded in the application.