feat: Update remotePatterns in next.config.js to allow images from localhost
This commit updates the `remotePatterns` array in the `next.config.js` file to allow images from the `localhost` domain. The `protocol` is changed to `http` and the `hostname` is set to `localhost`. This change improves the functionality of the application by allowing the loading of images from the local development environment.